fix: removed duplicate styles and integrated content block component with default and inverse versions
