+++ client/src/components/ContentBlock/styles.module.css
@@ -1,15 +1,3 @@
-@font-face {
-  font-family: Segoe Sans;
-  font-weight: 400;
-  src: url(../../../static/assets/fonts/Segoe-Sans-Display.ttf)
-    format("truetype");
-}
-@font-face {
-  font-family: Segoe Sans;
-  font-weight: 700;
-  src: url(../../../static/assets/fonts/Segoe-Sans-Display-Bold.ttf)
-    format("truetype");
-}
 .l-title h2 {
   position: relative;
   top: 0.0625rem;
@@ -139,16 +127,16 @@ a.l-btn:focus-visible {
 
 /* Common style ends */
 
-.section.section--content-block {
+.msf.content-block {
   padding: 1.5rem 0 2.375rem;
 }
 @media screen and (min-width: 991px) {
-  .section.section--content-block {
+  .msf.content-block {
     padding: 3rem 0;
   }
 }
 @media screen and (min-width: 1441px) {
-  .section.section--content-block {
+  .msf.content-block {
     padding: 3.3333333333vw 0;
   }
 }
