+++ client/src/pages/block/block.module.css
@@ -10,263 +10,15 @@
   src: url(../../../static/assets/fonts/Segoe-Sans-Display-Bold.ttf)
     format("truetype");
 }
-.l-title h2 {
-  position: relative;
-  top: 0.0625rem;
-  padding-bottom: clamp(0.75rem, 1.44509vw + 0.386921875rem, 1.6875rem);
-  margin-bottom: 0;
-  color: #0078d4;
-  font-family:
-    Segoe Sans,
-    sans-serif;
-  font-size: 2.25rem;
-  line-height: 122.222%;
-}
-[data-theme="dark"] .l-title h2 {
-  color: #4796e8;
-}
-@media screen and (min-width: 1540px) {
-  .l-title h2 {
-    font-size: 2.5vw;
-  }
-}
-@media screen and (min-width: 1320px) {
-  .l-title h2 {
-    top: 0;
-  }
-}
-@media screen and (min-width: 1441px) {
-  .l-title h2 {
-    padding-bottom: 1.875vw;
-  }
-}
-.l-title a.l-btn {
-  margin-top: 1.5rem;
-  font-family:
-    Segoe Sans,
-    sans-serif;
-}
-.l-title a.l-btn:hover {
-  text-decoration: none;
-}
-@media screen and (min-width: 1441px) {
-  .l-title a.l-btn {
-    margin-top: 1.6666666667vw;
-  }
-}
-.aspect-box {
-  position: relative;
-  display: block;
-  width: 100%;
-}
-.aspect-box canvas,
-.aspect-box iframe,
-.aspect-box img,
-.aspect-box svg,
-.aspect-box video {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-.aspect-box:before {
-  display: block;
-  padding-top: calc(var(--aspect-ratio) * 100%);
-  content: "";
-}
-a.l-btn {
-  display: inline-block;
-  min-width: 9.4375rem;
-  min-height: 1.5rem;
-  align-items: center;
-  justify-content: center;
-  padding: 0.5rem 0.75rem;
-  border-radius: 0.5rem;
-  background-color: #0078d4;
-  box-shadow: none;
-  color: #fff;
-  font-size: 1.25rem;
-  font-weight: 700;
-  line-height: 120%;
-  text-align: center;
-  text-decoration: none;
-  transition: all 0.3s ease-in-out;
-}
-@media screen and (min-width: 1540px) {
-  a.l-btn {
-    font-size: 1.3888888889vw;
-  }
-}
-@media screen and (min-width: 768px) {
-  a.l-btn {
-    display: inline-block;
-  }
-}
-@media screen and (min-width: 1540px) {
-  a.l-btn {
-    min-width: 10.4861111111vw;
-    min-height: 1.6666666667vw;
-    padding: 0.5555555556vw 0.8333333333vw;
-    border-radius: 0.5555555556vw;
-  }
-}
-a.l-btn:hover {
-  background-color: #104e90;
-  color: #fff;
-}
-a.l-btn:focus-visible {
-  outline: 0.0625rem solid #000;
-  outline-offset: 0.1875rem;
-}
-.msf .container {
-  max-width: 90rem;
-  padding: 0 1.5rem;
-  margin: 0 auto;
-}
-@media screen and (min-width: 1320px) {
-  .msf .container {
-    padding: 0 8.5rem;
-  }
-}
-@media screen and (min-width: 1441px) {
-  .msf .container {
-    max-width: 100vw;
-    padding: 0 9.4444444444vw;
-  }
-}
-/* Common style ends */
 
 .mainBlockContainer * {
   margin: 0;
   padding: 0;
+  font-family: Segoe Sans;
 }
 @media screen and (min-width: 1440px) {
   .mainBlockContainer {
     padding-top: 48px;
     padding-bottom: 48px;
-  }
-}
-.msf.card {
-  position: relative;
-  padding: 1.8125rem 2.1875rem;
-  border: 0.0625rem solid #d1d5db;
-  border-radius: 0.5rem;
-  background-color: #fff;
-  box-shadow: 0rem 0.125rem 0.5rem 0rem rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease-in-out;
-}
-.msf.card:hover {
-  background-color: rgba(0, 120, 212, 0.1);
-}
-[data-theme="dark"] .msf.card {
-  border-color: #2a2a2a;
-  background: rgba(0, 120, 212, 0.1);
-}
-[data-theme="dark"] .msf.card:hover {
-  background-color: #2a2a2a;
-}
-.msf.card a {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  font-size: 0;
-  inset: 0;
-}
-.msf.card .card__wrapper header {
-  display: flex;
-  align-items: center;
-  margin-bottom: 0.9375rem;
-  gap: 1.125rem;
-}
-@media screen and (min-width: 1441px) {
-  .msf.card .card__wrapper header {
-    gap: 1.25vw;
-  }
-}
-.msf.card .card__wrapper header figure {
-  max-width: 2.5rem;
-  margin: 0;
-}
-@media screen and (min-width: 1441px) {
-  .msf.card .card__wrapper header figure {
-    max-width: 2.7777777778vw;
-  }
-}
-.msf.card .card__wrapper header figure.aspect-box {
-  --aspect-ratio: 36/40;
-}
-.msf.card .card__wrapper header figure img {
-  width: 2.5rem;
-  height: 2.25rem;
-}
-@media screen and (min-width: 1441px) {
-  .msf.card .card__wrapper header figure img {
-    width: 2.7777777778vw;
-    height: 2.5vw;
-  }
-}
-.msf.card .card__wrapper header h2 {
-  position: relative;
-  top: 0.0625rem;
-  margin-bottom: 0;
-  color: #0078d4;
-  font-family:
-    Segoe Sans,
-    sans-serif;
-  font-size: 1.25rem;
-  line-height: 120%;
-}
-[data-theme="dark"] .msf.card .card__wrapper header h2 {
-  color: #5ea0ef;
-}
-@media screen and (min-width: 1540px) {
-  .msf.card .card__wrapper header h2 {
-    font-size: 1.3888888889vw;
-  }
-}
-.msf.card .card__wrapper article p {
-  color: #000;
-  font-family:
-    Segoe Sans,
-    sans-serif;
-  font-size: 1rem;
-  line-height: 137.5%;
-}
-[data-theme="dark"] .msf.card .card__wrapper article p {
-  color: #fff;
-}
-@media screen and (min-width: 1540px) {
-  .msf.card .card__wrapper article p {
-    font-size: 1.1111111111vw;
-  }
-}
-.section.section--card-list {
-  padding: 1.5rem 0 2.375rem;
-}
-@media screen and (min-width: 991px) {
-  .section.section--card-list {
-    padding: 3rem 0;
-  }
-}
-@media screen and (min-width: 1441px) {
-  .section.section--card-list {
-    padding: 3.3333333333vw 0;
-  }
-}
-.msf.card-list {
-  display: grid;
-  gap: 1.614375rem;
-}
-@media screen and (min-width: 768px) {
-  .msf.card-list {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-@media screen and (min-width: 991px) {
-  .msf.card-list {
-    gap: 2.4375rem 2.25rem;
-    grid-template-columns: repeat(3, 1fr);
   }
 }
